Trace: Update ti.trace.SysMin to be compatible with XDC 3.25
XDC 3.25 includes device support we will need, as well as support for CCS 5.4,
a key development environment.  Unfortunately, there is a small compatibility
break between XDC 3.24 and 3.25 for the [rare] modules that implement
xdc.runtime.ISystemSupport.  Unfortunately, ti.trace.SysMin is susceptible to
this break, and requires a slight change.

Prior to this commit, we must use XDC 3.24.  After this commit, we must use
XDC 3.25.