QNX IPC: Update MessageQCopy to Add Desc Field in Name Map
The BIOS-side NameMap module now sends a new desc field.
MessageQCopy, the module that receives the name map
messages, is updated to take this field into account.

As such, the callback that users of MessageQCopy can
register for notifications of remote core endpoint
creation is updated to give the desc field.  All users
of the MessageQCopy_registerNotify function must
be updated.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>