OMAP5 VirtQueue: Removed VirtQueue_postInitDone().
Looking at the Linux side trace, we see this message:

  omap-rproc omap-rproc.0: Dropping unknown message ffffff07

proving the Linux driver is not even processing this message.

Per Suman:
"This call was added as part of optimizing the boot-time for the Ducati image,
where in we starting loading at OPP100, and this message is used to trigger
releasing the OPP100 constraint.

You can remove it or keep it, the rproc message is discarding it anyway."

So, it was removed.

Signed-off-by: G Anthony <a0783926@ti.com>
Signed-off-by: Chris Ring <cring@ti.com>