Vayu: TableInit Module should be metaonly
The TableInit Module is incorrectly specified to be a target module, yet has
no target code, and no TableInit.c.

In addition to being more efficient as a metaonly module, this fixes a build
failure for Vayu platforms when using the BIOS custom build.  The custom build
attempts to find/build a .c source file for all target modules... which
doesn't exist for TableInit.