QNX IPC: WA For DSP Not Receiving Messages
There is an issue observed when testing HOST->DSP1
communication using Vayu Virtio where if more than one
message is in the mailbox FIFO, then the DSP only
picks up one message, and the interrupt callback
stops running.

To avoid this issue, the workaround is to make sure
that we only send one message at a time from HOST->DSP1.

A side affect of this is that messages will be dropped,
but since (for the most part), the contents of the mailbox
message are the same, we can get away with this for now.

This may be just a Vayu Virtio issue, and we will
attempt to revert the WA in the future.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>