IPC: Vayu - Fix GPTimer Values and Control Module AMMU Config
The GPTimer values in IpcCommon.cfg.xs were using the
old virtual addresses that started with 0xA8.  A
recent patch changed the virtual address mapping for
0x48000000 from 0xA8000000 to 0x68000000.  The
GPTimer base addresses missed being updated.

This patch updates the GPTimer values to the correct
virtual addresses.

Also, the control module entry in IpuAmmu.cfg is
updated to do a translation from 0x4a000000 to
0x6a000000, to take advantage of the existing
mapping of the 0x6a000000 to the 0x4a000000 registers.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>