QNX IPC: Remove EVE from MultiProc as it is Not Used Currently
If EVE is defined in the MultiProc, then when the IPU is
starting up, the interrupt module will try to access the EVE
mailboxes, which leads to a crash as we are not mapping
the EVE mailboxes.

This patch updates the MultiProc config to remove the EVEs
so that the EVE mailboxes are not accessed.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>