QNX IPC: Samples - Only Open SharedMemLib Once
A race condition in the rpmsg-rpc stress test could
result in one thread trying to access a NULL pointer
as there was no protection around the call to
initialize the sharedmemallocator library functions.

To avoid the race condition, move the initialization
of the sharedmemallocator functions to the main function,
to happen only once.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>