QNX IPC: Update Int Xbar Programming to Match BIOS-side IPC
There was a conflict with the MPU Int being used by QNX IPC
for the DSP Mailbox.

The BIOS-side IPC is using MPU INT 141 for HOST receiving
interrupts from IPU2, and MPU INT 136 for HOST receiving
interrupts from DSP1.

QNX IPC is updated to use the same as we know these interrupts
are available.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>