VirtQueue: VirtQueue_init() reads vring addr from resource table, checks MAR.
Previously, VirtQueue used hardcoded vring addresses in the code.  Now, these
are read from the Resource Table, allowing the host to update them at load
time.

However, it is still expected that the vring memory is non-cached, and that
the system configuration has set this up in the CFG file.

VirtQueue_init() asserts if the vring addresses are not in the resource table,
or if the MAR bits for the vring memory show it as being caceable.

This is done only for OMAPL138 and TCI6638 VirtQueue in this patch.

Signed-off-by: G Anthony <a0783926@ti.com>