IPC - Match Resource Table for Vayu Based on Device
Match the resource table for Vayu based on the Device
name instead of the Platform name.

Currently, if the platform simVayu is used, it fails
to find the resource table.  If the match is changed
to match the device name, then the simVayu platform,
which uses the Vayu device name, can compile.

Signed-off-by: Chris Ring <cring@ti.com>