OMAP4: Removed files specific to OMAP4 in qnx directory
This commit removes all source files that are specific to OMAP4,
which the product does not support. Note that there are still file
names, variable names, macros, etc. related to OMAP4 that will need
subsequent clean up.

Files in tiler and rpmsg-resmgr directories have been left in place
as we are discussing their removal in an on-going thread.

A minor change was needed in syslink_main.c as well in order to
eliminate its use of type 'u32' defined in rprcfmt.h, so that we can
remove the latter file, which is specific to the OMAP4 loader.

Signed-off-by: VW <vwan@ti.com>