QNX IPC: RPMSG-RPC - Don't Wait for Response During Destroy
When calling rpmsg_rpc_destroy, all existing connections
are disconnected so that the remote core may be reloaded
and stale connections are not left open.

When rpmsg_rpc_destroy is called, it is assumed that
the remote processor has already been stopped.  In this
case, there is no need to wait for a response from the
disconnect call.  Instead, the connection can immediately
be marked as disconnected and destroy can continue.

This patch modifies each remote core connection to have
a parameter that tells whether or not the destroy function
is currently executing.  Then, the disconnect call is
updated to check to see if destroy is being executed.  If
destroy is being executed, then it does not send the
message to the remote core and wait for a response.  Instead,
the connection is simply marked as disconnected.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>