QNX: Don't Disable Host->rproc Interrupts During Register
The register/unregister functions are clearing mailbox
messages and disabling the host->rproc interrupts.  When loading
multiple remote procs, this can cause an issue where the
remote proc's host->rproc interrupt remains disabled and the
remote proc is unable to receive interrupts.

This patch removes the code to avoid disabling the interrupt.