Vayu: Remove ti.platform.vayu.* pkgs
In the IPC 3.00 stream the tests/examples switched to
ti.platforms.evmDRA7XX platform instances, but we left
these ti.platform.vayu.* platform packages in the IPC product
to not disrupt existing users of them.

Users have had adequate time to move to the XDC-provided
platform instances, so we are removing them in the IPC 3.10
stream.