Resource: Add accessor to read the status field from a VDEV struct.
The Linux host will write a VIRTIO_CONFIG_* status into the remote processor
resource table upon completion of VDEV creation.

As this corresponds to priming of vring buffers for RPMSG type VDEVs, this is
a useful status for the slave to poll before it sends it's first RPMSG
to the host.

Other non-Linux HLOS must implement a similar status update mechanism.

Signed-off-by: G Anthony <a0783926@ti.com>