Linux: Remove KERNEL and DRM requirement
Remove the KERNEL and DRM dependency for OMAP54xx and
DRA7xx builds.  We'll build as many libs/tests as possible,
given the provided dependencies.

If the user doesn't supply KERNEL, we won't build the MmRpc
libs; if they don't supply DRM, we won't build tests that
depend on DRM... but we'll still build, for example,
MessageQ libs and tests.