GateMP: Introduce support for GateMP on QNX
To enable GateMP,throw the -g option when launching the IPC resource
manager.

On the BIOS side, DSP1 must be configured as the owner of SR0 and
the new GateMP.hostSupport config param must be set to true like
this:

     var GateMP = xdc.useModule('ti.sdo.ipc.GateMP');
     GateMP.hostSupport = true;

Note that it's initially only been validated on DRA7XX DSP1.

Signed-off-by: VW <vwan@ti.com>