QNX IPC: rpmsg_rpc - Don't Use Primary Offset in Translations
The primary offset for the params is not needed in the
translations. The translation code was incorrectly subtracting
the primary offset when calculating the location of the pointer
to be translated. The translation offset is the offset from the
start of the params data pointer, and not the offset from the
start of the params base. So, the params data pointer plus the
translation offset is all that is needed in order to get the
location of the translation.

Also, fix the reverse translations so that the correct
value is put back in the buffer.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>