Build: Align products.mak variables across OS's
The make variable to indicate a platform is now consistent
between BIOS, QNX and Linux, and only needs to be set once.

Note that this is a compatibility break for consumers (e.g. SDKs)
which automate the consumption/build of IPC.  This is consciously
being introduced in a new Major.minor release (3.20) to keep
from breaking 3.00 and 3.10 users.

This resolves SDOCM00103415.