SDOCM00103636: ping_rpmsg OMAP5 test broken
The OMAP5 and DRA7XX Linux kernels require slave-side code to
use RPMSG_NS_2_0-compatible calls to NameMap_register().  When
moving to the XDC-provided ti.platforms.sdp5430 platform, the
check for which platforms need to throw -DRPMSG_NS_2_0 broke.
As a result, the ping_rpmsg test, specifically for OMAP5, was
building incorrectly.