Remove ducati trace thread in QNX
The ducati trace thread was deemed redundant wrt. the IPC trace daemon,
and was stealing early trace when the -g option is thrown due to a longer
delay between the time when the slave is started and when the thread is
created. Hence it has been removed. This patch also adjusts the trace buffer size
in GateMPApp to match up with its resource entry. This addresses SDOCM00103794.

Signed-off-by: VW <vwan@ti.com>